Add vkcapture-show
This allows leveraging obs-vkcapture to get the content of Azahar. This works without OBS.  This will allow creating an oversized window and then capturing the content that overflows off the screen.

This means soon, this project can support gamescope by using "-S fill" as documented here https://github.com/sonic2kk/steamtinkerlaunch/wiki/GameScope

The layout will put the top screen in the center and then the bottom screen at the bottom, with a black bar at the top of equal size to the bottom screen to ensure that the top screen is centered.